fix(medications): tidy the supplement editor after review
Preserve a supplement's stored `dose_amount` when editing. The form pinned it
to 1 on every save, which would undo a dose set through another surface; the
dose is gated server side now, and this form does not offer the field for
supplements. New supplements still start at one dose.

Drop the dialog's controlled `open`/`onOpenChange` props: no consumer passes
them.

Silence the per-nutrient "Custom nutrient added" toast when the supplement
editor provisions rows on save. One free-text nutrient per row stacked N toasts
in front of the save toast the user was waiting on. Settings and the food
nutrient dialog keep theirs.

Key `MedicationNutrients` off the shared `FoodVariantNutrientField` union rather
than restating all seventeen field names, and fold `nutrientPickerUtils` into
`medicationUtils` now that it is two functions with one consumer.
