+++ SparkyFitnessFrontend/src/pages/Medications/NutrientPicker.tsx
@@ -10,7 +10,7 @@ import {
   collectClaimedNutrientNames,
   isNutrientOptionAlreadyAdded,
   type NutrientPick,
-} from './nutrientPickerUtils';
+} from './medicationUtils';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
